feat: extract summaries from Overview
